+++ src/kernel/alloc.h
@@ -2,10 +2,11 @@
 #define ALLOC_H
 
 #include "types.h"
+#include <stddef.h>
 
 void init_heap(void);
 
-void *malloc(u64 size);
+void *malloc(size_t size);
 void free(void *ptr);
 
 #endif
