+++ src/kernel/boot_info.c
@@ -1,16 +1,16 @@
 #include "boot_info.h"
 #include "stdio.h"
 
-u64 kernel_start;
-u64 kernel_end;
-u64 multiboot_start;
-u64 multiboot_end;
+size_t kernel_start;
+size_t kernel_end;
+size_t multiboot_start;
+size_t multiboot_end;
 
-u64 kernel_page_table_start;
-u64 kernel_page_table_end;
+size_t kernel_page_table_start;
+size_t kernel_page_table_end;
 
 memory_map_entry_t *memory_map;
-u64 memory_map_count;
+size_t memory_map_count;
 
 void init_boot_info(boot_info_t *boot_info)
 {
@@ -35,7 +35,7 @@ void init_boot_info(boot_info_t *boot_info)
             ptr += 8; // Jump over entry_size and entry_version
             memory_map = (memory_map_entry_t *)ptr;
             memory_map_count = (tag->size - 16) / sizeof(memory_map_entry_t);
-            for (u64 i = 0; i < memory_map_count; i++)
+            for (size_t i = 0; i < memory_map_count; i++)
             {
                 memory_map_entry_t *entry = memory_map + i;
                 if (entry->type == 1)
@@ -52,9 +52,9 @@ void init_boot_info(boot_info_t *boot_info)
             ptr += 3;
             elf_section_entry_t *entry = (elf_section_entry_t *)ptr;
             // printf("Kernel sections: \n");
-            u64 min = 0xffffffffffffffff;
-            u64 max = 0x0;
-            while ((u64)entry < (u64)tag + tag->size)
+            size_t min = 0xffffffffffffffff;
+            size_t max = 0x0;
+            while ((size_t)entry < (size_t)tag + tag->size)
             {
                 if (entry->type != 0) // skip unused (type 0x0)
                 {
@@ -71,8 +71,8 @@ void init_boot_info(boot_info_t *boot_info)
 
             kernel_start = min;
             kernel_end = max;
-            multiboot_start = (u64)boot_info;
-            multiboot_end = (u64)boot_info + boot_info->total_size;
+            multiboot_start = (size_t)boot_info;
+            multiboot_end = (size_t)boot_info + boot_info->total_size;
         }
 
         tag = (boot_tag_t *)((u8 *)tag + ((tag->size + 7) & ~7));
