+++ src/kernel/paging.h
@@ -2,6 +2,7 @@
 #define PAGING_H
 
 #include "types.h"
+#include <stddef.h>
 
 #define ENTRY_COUNT 512
 
@@ -12,10 +13,10 @@ typedef struct page_table
 
 page_table_t *level_4_table();
 
-u64 get_physaddr(void *virt_addr);
+size_t get_physaddr(void *virt_addr);
 
 void identity_map(void *virt_addr, u8 flags);
-void map_page(u64 phys_addr, void *virt_addr, u8 flags);
+void map_page(size_t phys_addr, void *virt_addr, u8 flags);
 void unmap_page(void *virt_addr);
 void *allocate_page();
